+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -130,7 +130,8 @@ export enum AdUnitTracking {
     AttemptToShowAd = 'ad_unit_attempt_to_show',
     FailedToInvalidate = 'ad_unit_failed_to_invalidate',
     MissedImpression = 'ad_unit_missed_impression',
-    MediationShowCall = 'ad_unit_mediation_show_call'
+    MediationShowCall = 'ad_unit_mediation_show_call',
+    RealMissedImpression = 'ad_unit_real_missed_impression'
 }
 
 type PTSEvent = AdmobMetric | BannerMetric | CachingMetric | ChinaMetric | VastMetric | MraidMetric | MiscellaneousMetric | LoadMetric | ProgrammaticTrackingError | OMMetric | TimingMetric | AUIMetric | AdUnitTracking;