+++ test/Unit/Ads/MissedImpressionManagerTest.ts
@@ -10,6 +10,7 @@ import { HttpKafka, KafkaCommonObjectType } from 'Core/Utilities/HttpKafka';
 import 'mocha';
 import * as sinon from 'sinon';
 import { TestFixtures } from 'TestHelpers/TestFixtures';
+import { ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingService';
 
 describe('MissedImpressionManagerTest', () => {
     let missedImpressionManager: MissedImpressionManager;
@@ -17,6 +18,7 @@ describe('MissedImpressionManagerTest', () => {
     let backend: Backend;
     let nativeBridge: NativeBridge;
     let core: ICoreApi;
+    let pts: ProgrammaticTrackingService;
     let kafkaSpy: any;
 
     beforeEach(() => {
@@ -25,7 +27,8 @@ describe('MissedImpressionManagerTest', () => {
         nativeBridge = TestFixtures.getNativeBridge(platform, backend);
         core = TestFixtures.getCoreApi(nativeBridge);
         core.Storage = new StorageApi(nativeBridge);
-        missedImpressionManager = new MissedImpressionManager(core);
+        pts = TestFixtures.getCoreModule(nativeBridge).ProgrammaticTrackingService;
+        missedImpressionManager = new MissedImpressionManager(core, pts, '', '');
         kafkaSpy = sinon.spy(HttpKafka, 'sendEvent');
     });
 
@@ -38,11 +41,14 @@ describe('MissedImpressionManagerTest', () => {
 
         assert.isTrue(kafkaSpy.calledOnce, 'missed impression event was not sent to httpkafka');
         assert.isTrue(kafkaSpy.calledWith('ads.sdk2.events.missedimpression.json', KafkaCommonObjectType.ANONYMOUS, { ordinal: 1 }), 'missed impression event arguments incorrect');
+
+        assert.isTrue((<sinon.SinonStub>pts.reportMetricEventWithTags).calledOnce, 'missed impression event was not sent to pts');
     });
 
     it('should not send events when other metadata is set', () => {
         core.Storage.onSet.trigger(StorageType[StorageType.PUBLIC], {'player': {'server_id': { 'value': 'test', 'ts': 123456789 }}});
 
         assert.isFalse(kafkaSpy.called, 'missed impression event was triggered for unrelated storage event');
+        assert.isFalse((<sinon.SinonStub>pts.reportMetricEventWithTags).called, 'missed impression event was triggered for unrelated storage event');
     });
 });